separate itensity corrections from slices #REF457
